Download AllCertificateRecordsCSVFormatv4 from the new two parts (V4a and V4b).

+++ fetch_csv_reports.sh
@@ -4,19 +4,33 @@ CURDIR=`pwd`
 TMPDIR=`mktemp -d`
 cd $TMPDIR
 
-wget -nv https://ccadb.my.salesforce-sites.com/ccadb/AllCertificateRecordsCSVFormatv4
-csvsort AllCertificateRecordsCSVFormatv4 > AllCertificateRecordsCSVFormatv4.sorted
-mv AllCertificateRecordsCSVFormatv4.sorted AllCertificateRecordsCSVFormatv4
+wget -nv -O AllCertificateRecordsCSVFormatv4 https://ccadb.my.salesforce-sites.com/ccadb/AllCertificateRecordsCSVFormatV4a
+if [ -s AllCertificateRecordsCSVFormatv4 ]; then
+  wget -nv https://ccadb.my.salesforce-sites.com/ccadb/AllCertificateRecordsCSVFormatV4b
+  if [ -s AllCertificateRecordsCSVFormatV4b ]; then
+    sed -i -e '$a\' AllCertificateRecordsCSVFormatv4
+    tail -n=+1 AllCertificateRecordsCSVFormatV4b >> AllCertificateRecordsCSVFormatv4
+    csvsort AllCertificateRecordsCSVFormatv4 > AllCertificateRecordsCSVFormatv4.sorted
+    mv AllCertificateRecordsCSVFormatv4.sorted AllCertificateRecordsCSVFormatv4
+  fi
+  rm -f AllCertificateRecordsCSVFormatV4b
+fi
 
 for i in $( seq 1994 `date +%Y` ); do
   wget -nv -O AllCertificatePEMsCSVFormat_NotBeforeYear_$i https://ccadb.my.salesforce-sites.com/ccadb/AllCertificatePEMsCSVFormat?NotBeforeYear=$i
-  csvsort AllCertificatePEMsCSVFormat_NotBeforeYear_$i > AllCertificatePEMsCSVFormat_NotBeforeYear_$i.sorted
-  mv AllCertificatePEMsCSVFormat_NotBeforeYear_$i.sorted AllCertificatePEMsCSVFormat_NotBeforeYear_$i
+  if [ -s AllCertificatePEMsCSVFormat_NotBeforeYear_$i ]; then
+    csvsort AllCertificatePEMsCSVFormat_NotBeforeYear_$i > AllCertificatePEMsCSVFormat_NotBeforeYear_$i.sorted
+    mv AllCertificatePEMsCSVFormat_NotBeforeYear_$i.sorted AllCertificatePEMsCSVFormat_NotBeforeYear_$i
+  else
+    rm -f AllCertificatePEMsCSVFormat_NotBeforeYear_$i
+  fi
 done
 
 cd $CURDIR
 mkdir -p data
-mv $TMPDIR/AllCertificateRecordsCSVFormatv4 $CURDIR/data
+if [ -s $TMPDIR/AllCertificateRecordsCSVFormatv4 ]; then
+  mv $TMPDIR/AllCertificateRecordsCSVFormatv4 $CURDIR/data
+fi
 mv $TMPDIR/* $CURDIR/cmd/ski_spki/data
 rmdir $TMPDIR
